+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - New `error/3` callback for process manager and deprecated `error/4` ([#124](https://github.com/commanded/commanded/pull/124))
 - Router support for identity prefix function.
 - Retry command execution on concurrency error ([#132](https://github.com/commanded/commanded/pull/132)).
+- Event handler `error/3` callback ([#133](https://github.com/commanded/commanded/pull/133)).
 
 ## v0.15.1